Update to latest

+++ bmtk/utils/reports/compartment/compartment_writer.py
@@ -27,13 +27,15 @@ class PopulationWriterv01(CompartmentWriterABC, CompartmentReader):
     parallel mode. For better performance use the CellVarRecorderParrallel class instead.
     """
     class DataTable(object):
-        """A small struct to keep track of different */data (and buffer) tables"""
+        """A small struct to keep track of different data (and buffer) tables
+        """
         def __init__(self, var_name):
             self.var_name = var_name
             # If buffering data, buffer_block will be an in-memory array and will write to data_block during when
             # filled. If not buffering buffer_block is an hdf5 dataset and data_block is ignored
             self.data_block = None
             self.buffer_block = None
+            self.block_window = (0.0, 10e20)
 
     def __init__(self, parent, population, variable=None, units=None, tstart=0.0, tstop=1.0, dt=0.01, n_steps=None,
                  buffer_size=0, **kwargs):
@@ -64,7 +66,7 @@ class PopulationWriterv01(CompartmentWriterABC, CompartmentReader):
         self._buffer_size = buffer_size
         self._buffer_data = buffer_size > 0
         self._data_block = self.DataTable(self._variable)
-        self._last_save_indx = 0  # for buffering, used to keep track of last timestep data was saved to disk
+        # self._last_save_indx = 0  # for buffering, used to keep track of last timestep data was saved to disk
 
         self._buffer_block_size = 0
         self._total_steps = 0
@@ -173,7 +175,7 @@ class PopulationWriterv01(CompartmentWriterABC, CompartmentReader):
         var_grp = base_grp.create_group('mapping')
         var_grp.create_dataset('node_ids', shape=(self._n_gids_all,), dtype=np.uint)
         var_grp.create_dataset('element_ids', shape=(self._n_segments_all,), dtype=np.uint)
-        var_grp.create_dataset('element_pos', shape=(self._n_segments_all,), dtype=np.float)
+        var_grp.create_dataset('element_pos', shape=(self._n_segments_all,), dtype=float)
         var_grp.create_dataset('index_pointer', shape=(self._n_gids_all+1,), dtype=np.uint64)
         var_grp.create_dataset('time', data=[self.tstart(), self.tstop(), self.dt()])
         for k, v in self._element_data.items():
@@ -187,6 +189,7 @@ class PopulationWriterv01(CompartmentWriterABC, CompartmentReader):
             var_grp[k][self._seg_offset_beg:self._seg_offset_end] = v
 
         self._total_steps = n_steps
+        self._buffer_size = np.min((self._total_steps, self._buffer_size))
         self._buffer_block_size = self._buffer_size
 
         if not self._buffer_data:
@@ -195,23 +198,28 @@ class PopulationWriterv01(CompartmentWriterABC, CompartmentReader):
             for gid, gid_offset in self._gid_map.items():
                 self._gid_map[gid] = (gid_offset[0] + self._seg_offset_beg, gid_offset[1] + self._seg_offset_beg)
 
-        
         if self._buffer_data:
             # Set up in-memory block to buffer recorded variables before writing to the dataset
-            self._data_block.buffer_block = np.zeros((self._buffer_size, self._n_segments_local), dtype=np.float)
+            self._data_block.buffer_block = np.zeros((self._buffer_size, self._n_segments_local), dtype=float)
 
             self._data_block.data_block = base_grp.create_dataset('data', shape=(self.n_steps(), self._n_segments_all),
-                                                                  dtype=np.float, chunks=True)
+                                                                  dtype=float, chunks=True)
             if self._variable is not None:
                 self._data_block.data_block.attrs['variable'] = self._variable
 
             if self._units is not None:
                 self._data_block.data_block.attrs['units'] = self._units
 
+            self._data_block.block_window = (0, self._buffer_block_size)
+
         else:
             # Since we are not buffering data, we just write directly to the on-disk dataset
-            self._data_block.buffer_block = base_grp.create_dataset('data', shape=(self.n_steps(), self._n_segments_all),
-                                                               dtype=np.float, chunks=True)
+            self._data_block.buffer_block = base_grp.create_dataset(
+                'data',
+                shape=(self.n_steps(), self._n_segments_all),
+                dtype=float,
+                chunks=True
+            )
             if self._variable is not None:
                 self._data_block.buffer_block.attrs['variable'] = self._variable
 
@@ -219,6 +227,11 @@ class PopulationWriterv01(CompartmentWriterABC, CompartmentReader):
                 self._data_block.buffer_block.attrs['units'] = self._units
 
         self._is_initialized = True
+
+    def _reset_buffer_window(self, tstep):
+        blk_beg = int(tstep/self._buffer_size)*self._buffer_size
+        blk_end = blk_beg + self._buffer_size
+        self._data_block.block_window = (blk_beg, blk_end)
 
     def record_cell(self, node_id, vals, tstep, population=None):
         """Record cell parameters.
@@ -231,8 +244,19 @@ class PopulationWriterv01(CompartmentWriterABC, CompartmentReader):
         self.initialize()
         gid_beg, gid_end = self._gid_map[node_id]
         buffer_block = self._data_block.buffer_block
-        update_index = (tstep - self._last_save_indx)
-        buffer_block[update_index, gid_beg:gid_end] = vals
+
+        if not self._buffer_data:
+            buffer_block[tstep, gid_beg:gid_end] = vals
+
+        elif self._data_block.block_window[0] <= tstep < self._data_block.block_window[1]:
+            update_index = tstep - self._data_block.block_window[0]
+            buffer_block[update_index, gid_beg:gid_end] = vals
+
+        else:
+            self.flush()
+            self._reset_buffer_window(tstep)
+            update_index = tstep - self._data_block.block_window[0]
+            buffer_block[update_index, gid_beg:gid_end] = vals
 
     def record_cell_block(self, node_id, vals, beg_step, end_step, population=None):
         """Save cell parameters one block at a time
@@ -246,23 +270,24 @@ class PopulationWriterv01(CompartmentWriterABC, CompartmentReader):
         buffer_block = self._data_block.buffer_block
         if isinstance(vals, list) or vals.ndim == 1:
             buffer_block[:, gid_beg] = vals
-            #buffer_block[beg_step:end_step, gid_beg:gid_end] = vals
+            # buffer_block[beg_step:end_step, gid_beg:gid_end] = vals
         else:
             buffer_block[:, gid_beg:gid_end] = vals
 
     def flush(self):
         """Move data from memory to dataset"""
         if self._buffer_data:
-            blk_beg = self._last_save_indx
-            blk_end = blk_beg + self._buffer_block_size
+            # blk_beg = self._last_save_indx
+            # blk_end = blk_beg + self._buffer_block_size
+            blk_beg = self._data_block.block_window[0]
+            blk_end = self._data_block.block_window[1]
             if blk_end > self._total_steps:
                 # Need to handle the case that simulation doesn't end on a block step
                 blk_end = blk_beg + self._total_steps - blk_beg
 
             block_size = blk_end - blk_beg
-            self._last_save_indx += block_size
-
             self._data_block.data_block[blk_beg:blk_end, :] = self._data_block.buffer_block[:block_size, :]
+            self._reset_buffer_window(blk_end+1)
 
     def close(self):
         # Let the parent take care of this
@@ -298,7 +323,7 @@ class CompartmentWriterv01(CompartmentWriterABC):
         self._final_fpath = file_path  # name of file being writen too.
         self._cache_dir = cache_dir or os.path.dirname(os.path.abspath(file_path))  # used for mulitple ranks
         self._base_name = os.path.basename(file_path)  # make sure file names don't clash if there are multiple reports
-        self._interm_fpath = self._get_iterm_fpath() # In certain cases (parallelized simulation) split the final file by rank.
+        self._interm_fpath = self._get_iterm_fpath()  # In certain cases (parallelized simulation) split the final file by rank.
 
     def _get_iterm_fpath(self):
         if self._mpi_size > 1:
@@ -375,7 +400,10 @@ class CompartmentWriterv01(CompartmentWriterABC):
     def close(self):
         for pop_grp in self._pop_tables.values():
             pop_grp.close()
-        self._h5_handle.close()
+
+        if self._h5_handle is not None:
+            self._h5_handle.close()
+
         if self._mpi_size > 1:
             self.merge()
 
@@ -383,7 +411,8 @@ class CompartmentWriterv01(CompartmentWriterABC):
         barrier()
         if self._mpi_size > 1 and self._mpi_rank == 0:
             h5final = h5py.File(self._final_fpath, 'w')
-            tmp_reports = [CompartmentReader(name) for name in self.temp_files]
+
+            tmp_reports = [CompartmentReader(name) for name in self.temp_files if os.path.exists(name)]
             populations = set()
             for r in tmp_reports:
                 populations.update(r.populations)
@@ -428,7 +457,7 @@ class CompartmentWriterv01(CompartmentWriterABC):
                 if times is not None and len(times) > 0:
                     mapping_grp.create_dataset('time', data=times)
                 element_id_ds = mapping_grp.create_dataset('element_ids', shape=(total_seg_count,), dtype=np.uint)
-                el_pos_ds = mapping_grp.create_dataset('element_pos', shape=(total_seg_count,), dtype=np.float)
+                el_pos_ds = mapping_grp.create_dataset('element_pos', shape=(total_seg_count,), dtype=float)
                 gids_ds = mapping_grp.create_dataset('node_ids', shape=(total_gid_count,), dtype=np.uint)
                 index_pointer_ds = mapping_grp.create_dataset('index_pointer', shape=(total_gid_count + 1,),
                                                               dtype=np.uint)
@@ -468,7 +497,7 @@ class CompartmentWriterv01(CompartmentWriterABC):
                 # combine the /var/data datasets
                 data_name = '/report/{}/data'.format(pop)
                 # data_name = '/{}/data'.format(var_name)
-                var_data = h5final.create_dataset(data_name, shape=(n_steps, total_seg_count), dtype=np.float)
+                var_data = h5final.create_dataset(data_name, shape=(n_steps, total_seg_count), dtype=float)
                 # var_data.attrs['variable_name'] = var_name
                 i = 0
                 for rpt in tmp_reports:
@@ -487,7 +516,8 @@ class CompartmentWriterv01(CompartmentWriterABC):
                     var_data.attrs['units'] = units
 
             for tmp_file in self.temp_files:
-                os.remove(tmp_file)
+                if os.path.exists(tmp_file):
+                    os.remove(tmp_file)
         barrier()
 
     def _build_or_fetch_pop(self, population):
